+++ pkg/glib/glib.go
@@ -11,11 +11,12 @@ import (
 	password "github.com/sethvargo/go-password/password"
 	"golang.org/x/oauth2/google"
 	admin "google.golang.org/api/admin/directory/v1"
+	groupssettings "google.golang.org/api/groupssettings/v1"
 	"google.golang.org/api/option"
 )
 
-// CreateDirectoryService() creates a client for communicating with Google APIs,
-// returns an Admin SDK Directory service object authorized with.
+// NewDirectoryService() creates a client for communicating with Google Directory API,
+// returns a service object authorized to perform actions in Gsuite.
 func NewDirectoryService(clientSecretFile string, impersonatedUserEmail string) (*admin.Service, error) {
 	ctx := context.Background()
 
@@ -24,7 +25,7 @@ func NewDirectoryService(clientSecretFile string, impersonatedUserEmail string) 
 		return nil, fmt.Errorf("ReadFile(clientSecretFile): %v", err)
 	}
 
-	config, err := google.JWTConfigFromJSON(jsonCredentials, admin.AdminDirectoryUserScope, admin.AdminDirectoryGroupScope, admin.AdminDirectoryGroupMemberScope, admin.AdminDirectoryOrgunitScope)
+	config, err := google.JWTConfigFromJSON(jsonCredentials, admin.AdminDirectoryUserScope, admin.AdminDirectoryGroupScope, admin.AdminDirectoryGroupMemberScope, admin.AdminDirectoryOrgunitScope, admin.AdminDirectoryResourceCalendarScope)
 	if err != nil {
 		return nil, fmt.Errorf("JWTConfigFromJSON: %v", err)
 	}
@@ -33,6 +34,31 @@ func NewDirectoryService(clientSecretFile string, impersonatedUserEmail string) 
 	ts := config.TokenSource(ctx)
 
 	srv, err := admin.NewService(ctx, option.WithTokenSource(ts))
+	if err != nil {
+		return nil, fmt.Errorf("NewService: %v", err)
+	}
+	return srv, nil
+}
+
+// NewGroupsService() creates a client for communicating with Google Groupssettings API,
+// returns a service object authorized to perform actions in Gsuite.
+func NewGroupsService(clientSecretFile string, impersonatedUserEmail string) (*groupssettings.Service, error) {
+	ctx := context.Background()
+
+	jsonCredentials, err := ioutil.ReadFile(clientSecretFile)
+	if err != nil {
+		return nil, fmt.Errorf("ReadFile(clientSecretFile): %v", err)
+	}
+
+	config, err := google.JWTConfigFromJSON(jsonCredentials, groupssettings.AppsGroupsSettingsScope)
+	if err != nil {
+		return nil, fmt.Errorf("JWTConfigFromJSON: %v", err)
+	}
+	config.Subject = impersonatedUserEmail
+
+	ts := config.TokenSource(ctx)
+
+	srv, err := groupssettings.NewService(ctx, option.WithTokenSource(ts))
 	if err != nil {
 		return nil, fmt.Errorf("NewService: %v", err)
 	}
@@ -50,6 +76,7 @@ func GetListOfUsers(srv admin.Service) ([]*admin.User, error) {
 		log.Fatalf("Unable to retrieve users in domain: %v", err)
 		return nil, err
 	}
+
 	return request.Users, nil
 }
 
@@ -76,7 +103,7 @@ func CreateUser(srv admin.Service, user *config.UserConfig) error {
 		log.Fatalf("Unable to generate password: %v", err)
 		return err
 	}
-	newUser := createGSuiteUserFromConfig(user)
+	newUser := createGSuiteUserFromConfig(srv, user)
 	newUser.Password = pass
 	newUser.ChangePasswordAtNextLogin = true
 
@@ -100,7 +127,7 @@ func DeleteUser(srv admin.Service, user *admin.User) error {
 
 // UpdateUser updates the remote user with config
 func UpdateUser(srv admin.Service, user *config.UserConfig) error {
-	updatedUser := createGSuiteUserFromConfig(user)
+	updatedUser := createGSuiteUserFromConfig(srv, user)
 	_, err := srv.Users.Update(user.PrimaryEmail, updatedUser).Do()
 	if err != nil {
 		log.Fatalf("Unable to update a user: %v", err)
@@ -110,7 +137,7 @@ func UpdateUser(srv admin.Service, user *config.UserConfig) error {
 }
 
 // createGSuiteUserFromConfig converts a ConfigUser to (GSuite) admin.User
-func createGSuiteUserFromConfig(user *config.UserConfig) *admin.User {
+func createGSuiteUserFromConfig(srv admin.Service, user *config.UserConfig) *admin.User {
 	googleUser := &admin.User{
 		Name: &admin.UserName{
 			GivenName:  user.FirstName,
@@ -120,8 +147,44 @@ func createGSuiteUserFromConfig(user *config.UserConfig) *admin.User {
 		OrgUnitPath:  user.OrgUnitPath,
 	}
 
+	if len(user.Aliases) > 0 {
+		// TODO: fix it, doesnt add aliases
+		for _, alias := range user.Aliases {
+			googleUser.Aliases = append(googleUser.Aliases, string(alias))
+		}
+	}
+
+	if len(user.Phones) > 0 {
+		phNums := []admin.UserPhone{}
+		for _, phone := range user.Phones {
+			phNum := admin.UserPhone{
+				Value: phone,
+				Type:  "home",
+			}
+			phNums = append(phNums, phNum)
+		}
+		googleUser.Phones = phNums
+	}
+
+	if user.Address != "" {
+		addr := &[]admin.UserAddress{
+			{
+				Formatted: user.Address,
+				Type:      "home",
+			},
+		}
+		googleUser.Addresses = addr
+	}
+
+	if user.RecoveryEmail != "" {
+		googleUser.RecoveryEmail = user.RecoveryEmail
+	}
+
+	if user.RecoveryPhone != "" {
+		googleUser.RecoveryPhone = user.RecoveryPhone
+	}
+
 	if user.SecondaryEmail != "" {
-		googleUser.RecoveryEmail = user.SecondaryEmail
 		workEm := &[]admin.UserEmail{
 			{
 				Address: user.SecondaryEmail,
@@ -131,7 +194,104 @@ func createGSuiteUserFromConfig(user *config.UserConfig) *admin.User {
 		googleUser.Emails = workEm
 	}
 
+	if user.Employee != (config.EmployeeConfig{}) {
+		uOrg := &[]admin.UserOrganization{
+			{
+				Department:  user.Employee.Department,
+				Title:       user.Employee.JobTitle,
+				CostCenter:  user.Employee.CostCenter,
+				Description: user.Employee.Type,
+			},
+		}
+		googleUser.Organizations = uOrg
+		rel := &[]admin.UserRelation{
+			{
+				Value: user.Employee.ManagerEmail,
+				Type:  "manager",
+			},
+		}
+		googleUser.Relations = rel
+
+	}
+
+	if user.Location != (config.LocationConfig{}) {
+		//BuildingHandling(srv, user.Location)
+		loc := &[]admin.UserLocation{
+			{
+				BuildingId:   user.Location.Building,
+				FloorName:    user.Location.Floor,
+				FloorSection: user.Location.FloorSection,
+				Type:         "desk",
+			},
+		}
+		//googleUser.Locations = loc
+		fmt.Println(loc)
+
+	}
+
 	return googleUser
+}
+
+// createConfigUserFromGSuite converts a (GSuite) admin.User to ConfigUser
+func CreateConfigUserFromGSuite(googleUser *admin.User) config.UserConfig {
+	// get emails
+	primaryEmail, secondaryEmail := GetUserEmails(googleUser)
+
+	configUser := config.UserConfig{
+		FirstName:      googleUser.Name.GivenName,
+		LastName:       googleUser.Name.FamilyName,
+		PrimaryEmail:   primaryEmail,
+		SecondaryEmail: secondaryEmail,
+		OrgUnitPath:    googleUser.OrgUnitPath,
+		RecoveryPhone:  googleUser.RecoveryPhone,
+		RecoveryEmail:  googleUser.RecoveryEmail,
+	}
+
+	if len(googleUser.Aliases) > 0 {
+		for _, alias := range googleUser.Aliases {
+			configUser.Aliases = append(configUser.Aliases, string(alias))
+		}
+	}
+
+	if googleUser.Phones != nil {
+		for _, phone := range googleUser.Phones.([]interface{}) {
+			configUser.Phones = append(configUser.Phones, fmt.Sprint(phone.(map[string]interface{})["value"]))
+		}
+	}
+
+	if googleUser.Organizations != nil {
+		for _, org := range googleUser.Organizations.([]interface{}) {
+			configUser.Employee.Department = fmt.Sprint(org.(map[string]interface{})["department"])
+			configUser.Employee.JobTitle = fmt.Sprint(org.(map[string]interface{})["title"])
+			configUser.Employee.Type = fmt.Sprint(org.(map[string]interface{})["description"])
+			configUser.Employee.CostCenter = fmt.Sprint(org.(map[string]interface{})["costCenter"])
+		}
+	}
+
+	if googleUser.Relations != nil {
+		for _, rel := range googleUser.Relations.([]interface{}) {
+			if rel.(map[string]interface{})["type"] == "manager" {
+				configUser.Employee.ManagerEmail = fmt.Sprint(rel.(map[string]interface{})["value"])
+			}
+		}
+	}
+
+	if googleUser.Locations != nil {
+		for _, loc := range googleUser.Locations.([]interface{}) {
+			configUser.Location.Building = fmt.Sprint(loc.(map[string]interface{})["buildingId"])
+			configUser.Location.Floor = fmt.Sprint(loc.(map[string]interface{})["floorName"])
+			configUser.Location.FloorSection = fmt.Sprint(loc.(map[string]interface{})["floorSection"])
+		}
+	}
+	if googleUser.Addresses != nil {
+		for _, addr := range googleUser.Addresses.([]interface{}) {
+			if addr.(map[string]interface{})["type"] == "home" {
+				configUser.Address = fmt.Sprint(addr.(map[string]interface{})["formatted"])
+			}
+		}
+	}
+
+	return configUser
 }
 
 //----------------------------------------//
@@ -148,9 +308,19 @@ func GetListOfGroups(srv *admin.Service) ([]*admin.Group, error) {
 	return request.Groups, nil
 }
 
+// GetSettingOfGroup returns a group settings object from the API
+func GetSettingOfGroup(srv *groupssettings.Service, groupId string) (*groupssettings.Groups, error) {
+	request, err := srv.Groups.Get(groupId).Do()
+	if err != nil {
+		log.Fatalf("Unable to retrieve group settings: %v", err)
+		return nil, err
+	}
+	return request, nil
+}
+
 // CreateGroup creates a new group in GSuite via their API
 func CreateGroup(srv admin.Service, group *config.GroupConfig) error {
-	newGroup := createGSuiteGroupFromConfig(group)
+	newGroup, groupSettings := createGSuiteGroupFromConfig(group)
 	_, err := srv.Groups.Insert(newGroup).Do()
 	if err != nil {
 		log.Fatalf("Unable to create a group: %v", err)
@@ -160,6 +330,8 @@ func CreateGroup(srv admin.Service, group *config.GroupConfig) error {
 	for _, member := range group.Members {
 		AddNewMember(srv, newGroup.Email, &member)
 	}
+	fmt.Println(groupSettings)
+
 	return nil
 }
 
@@ -175,17 +347,18 @@ func DeleteGroup(srv admin.Service, group *admin.Group) error {
 
 // UpdateGroup updates the remote group with config
 func UpdateGroup(srv admin.Service, group *config.GroupConfig) error {
-	updatedGroup := createGSuiteGroupFromConfig(group)
+	updatedGroup, groupSettings := createGSuiteGroupFromConfig(group)
 	_, err := srv.Groups.Update(group.Email, updatedGroup).Do()
 	if err != nil {
 		log.Fatalf("Unable to update a group: %v", err)
 		return err
 	}
+	fmt.Println(groupSettings)
 	return nil
 }
 
 // createGSuiteGroupFromConfig converts a ConfigGroup to (GSuite) admin.Group
-func createGSuiteGroupFromConfig(group *config.GroupConfig) *admin.Group {
+func createGSuiteGroupFromConfig(group *config.GroupConfig) (*admin.Group, *groupssettings.Groups) {
 	googleGroup := &admin.Group{
 		Name:  group.Name,
 		Email: group.Email,
@@ -193,7 +366,17 @@ func createGSuiteGroupFromConfig(group *config.GroupConfig) *admin.Group {
 	if group.Description != "" {
 		googleGroup.Description = group.Description
 	}
-	return googleGroup
+
+	groupSettings := &groupssettings.Groups{
+		WhoCanContactOwner:   group.WhoCanContactOwner,
+		WhoCanViewMembership: group.WhoCanViewMembership,
+		WhoCanApproveMembers: group.WhoCanApproveMembers,
+		WhoCanPostMessage:    group.WhoCanPostMessage,
+		WhoCanJoin:           group.WhoCanJoin,
+		AllowExternalMembers: group.AllowExternalMembers,
+	}
+
+	return googleGroup, groupSettings
 }
 
 //----------------------------------------//
@@ -336,4 +519,44 @@ func createGSuiteOUFromConfig(ou *config.OrgUnitConfig) *admin.OrgUnit {
 	}
 
 	return googleOU
+}
+
+func BuildingHandling(srv admin.Service, location config.LocationConfig) error {
+	request, err := srv.Resources.Buildings.List("my_customer").Do()
+	if err != nil {
+		log.Fatalf("Unable to retrieve Buildings in domain: %v", err)
+		return err
+	}
+
+	found := false
+	for _, building := range request.Buildings {
+		fmt.Println(building)
+		if location.Building == building.BuildingId {
+			found = true
+			// update with floor & floor section
+			//for _, floor := range building.FloorNames {
+
+			//}
+
+			break
+		}
+	}
+	if !found {
+		// (r *ResourcesBuildingsService) Insert(customer string, building *Building)
+		newBuilding := &admin.Building{
+			BuildingId:   location.Building,
+			BuildingName: location.Building,
+			FloorNames:   []string{location.Floor},
+			//FloorSection: location.FloorSection,
+			//ResourceName: fmt.Sprintf("%s, %s, %s", location.Building, location.Floor, location.FloorSection),
+			//ResourceId:   location.Building,
+		}
+		_, err := srv.Resources.Buildings.Insert("my_customer", newBuilding).Do()
+		if err != nil {
+			log.Fatalf("Unable to create Building in domain: %v", err)
+			return err
+		}
+	}
+
+	return nil
 }